+++ src/cpx_io/cpx_system/cpx_ap/builder/apdd_information_builder.py
@@ -10,24 +10,28 @@ from cpx_io.utils.logging import Logging
 class Variant:
     """Variant dataclass"""
 
+    channel_group_ids: List[int]
     description: str
     name: str
     parameter_group_ids: List[int]
+    profile: List[int]
     variant_identification: dict
 
 
 def build_variant(variant_dict):
     """Builds one Variant"""
     return Variant(
+        variant_dict.get("ChannelGroupIds"),
         variant_dict.get("Description"),
         variant_dict.get("Name"),
         variant_dict.get("ParameterGroupIds"),
+        variant_dict.get("Profile"),
         variant_dict.get("VariantIdentification"),
     )
 
 
-def build_apdd_information(apdd, module_code):
-    """Builds one ApddInformation"""
+def build_actual_variant(apdd, module_code):
+    """Build variant and return the correct one according to module_code"""
     variant_list = [build_variant(d) for d in apdd["Variants"]["VariantList"]]
     Logging.logger.debug(f"Variants: {variant_list}")
     variant = next(
@@ -41,6 +45,11 @@ def build_apdd_information(apdd, module_code):
     if variant is None:
         raise IndexError(f"Could not find variant for ModuleCode {module_code}")
     Logging.logger.debug(f"Module Variant: {variant}")
+    return variant
+
+
+def build_apdd_information(apdd, variant):
+    """Builds one ApddInformation"""
 
     product_category = apdd["Variants"]["DeviceIdentification"]["ProductCategory"]
     product_family = apdd["Variants"]["DeviceIdentification"]["ProductFamily"]